Reorganized directory structure
In preparation for transformation to a library
